Fixed building and plugins on windows.  Need to test

+++ EagleLib/RuntimeObjectSystem/IObject.h
@@ -295,7 +295,7 @@ struct IObject
     IObject() : _isRuntimeDelete(false) {}
     virtual ~IObject()
     {
-        for(int i = 0; i < notifiers.size(); ++i)
+        for(size_t i = 0; i < notifiers.size(); ++i)
         {
             notifiers[i]->updateObject(nullptr);
         }
@@ -351,7 +351,7 @@ struct IObject
     }
     void updateNotifiers()
     {
-        for(int i = 0; i < notifiers.size(); ++i)
+        for(size_t i = 0; i < notifiers.size(); ++i)
         {
             notifiers[i]->updateObject(this);
         }
